+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from "./execute.ts";
+export * from "./mod.ts";
 export { HCEval } from "./execute/hc-eval.ts";
 export * from "./frames.ts";
 export * from "./maml.ts";